A Class to Contain PharmacoGenomic datasets together with their curations

The PharmacoSet (pSet) class was developed to contain and organise large PharmacoGenomic datasets, and aid in their metanalysis. It was designed primarily to allow bioinformaticians and biologists to work with data at the level of genes, drugs and cell lines, providing a more naturally intuitive interface and simplifying analyses between several datasets. As such, it was designed to be flexible enough to hold datasets of two different natures while providing a common interface. The class can accomidate datasets containing both drug dose response data, as well as datasets contaning genetic profiles of cell lines pre and post treatement with compounds, known respecitively as sensitivity and perturbation datasets.

Arguments

object

A PharmacoSet object

mDataType

A character with the type of molecular data to return/update

value

A replacement value

Value

An object of the PharmacoSet class

Slots

annotation

A list of annotation data about the PharmacoSet, including the $name and the session information for how the object was creating, detailing the exact versions of R and all the packages used

molecularProfiles

A list containing SummarizedExperiment type object for holding data for RNA, DNA, SNP and CNV measurements, with associated fData and pData containing the row and column metadata

sample

A data.frame containing the annotations for all the cell lines profiled in the data set, across all data types

treatment

A data.frame containg the annotations for all the drugs profiled in the data set, across all data types

treatmentResponse

A list containing all the data for the sensitivity experiments, including $info, a data.frame containing the experimental info,$raw a 3D array containing raw data, $profiles, a data.frame containing sensitivity profiles statistics, and $n, a data.frame detailing the number of experiments for each cell-drug pair

perturbation

A list containting $n, a data.frame summarizing the available perturbation data,

curation

A list containing mappings for $treatment, cell, tissue names used in the data set to universal identifiers used between different PharmacoSet objects

datasetType

A character string of 'sensitivity', 'perturbation', or both detailing what type of data can be found in the PharmacoSet, for proper processing of the data