Make a CoreSet with the updated class structure

New implementation of the CoreSet constructor to support MAE and TRE. This constructor will be swapped with the original CoreSet constructor as part of an overhaul of the CoreSet class structure.

Usage

PharmacoSet2(
  name = "emptySet",
  treatment = data.frame(),
  sample = data.frame(),
  molecularProfiles = MultiAssayExperiment(),
  treatmentResponse = TreatmentResponseExperiment(),
  perturbation = list(),
  curation = list(sample = data.frame(), treatment = data.frame(), tissue = data.frame()),
  datasetType = "sensitivity"
)

Arguments

name

A character(1) vector with the PharmacoSet objects name.

treatment

A data.frame with treatment level metadata. Treatments in a PharmacoSet represent pharmaceutical compounds.

sample

A data.frame with sample level metadata for the union of samples in treatmentResponse and molecularProfiles. Samples in a PharmacoSet represent cancer cell-lines.

molecularProfiles

A MultiAssayExperiment containing one SummarizedExperiment object for each molecular data type.

treatmentResponse

A LongTable or LongTableDataMapper object containing all treatment response data associated with the PharmacoSet object.

perturbation

A deprecated slot in a PharmacoSet object included for backwards compatibility. This may be removed in future releases.

curation

This class requires an additional curation item, tissue, which maps from published to standardized tissue idenifiers.

datasetType

A deprecated slot in a PharmacoSet object included for backwards compatibility. This may be removed in future releases.

Value

A CoreSet object storing standardized and curated treatment response and multiomic profile data associated with a given publication.

Examples

data(CCLEsmall)
CCLEsmall
#> <PharmacoSet>
#> Name:  CCLE 
#> Date Created:  Fri Nov  6 14:00:53 2015 
#> Number of samples:  1061 
#> Molecular profiles:
#> RNA :
#>   Dim: 50, 1036 
#> RNAseq :
#>   Dim: 50, 935 
#> mutation :
#>   Dim: 1667, 1052 
#> CNV :
#>   Dim: 50, 744 
#> Treatment response:
#> Drug pertubation:
#>    Please look at pertNumber(cSet) to determine number of experiments  for each drug-sample combination.
#> Drug sensitivity:
#>    Number of Experiments:  11670 
#>    Please look at sensNumber(cSet) to determine number of  experiments for each drug-sample combination.