Computes Loewe Null References

Predict the response of a treatment combination under the Loewe additive null assumption.

Usage

computeLoewe(
  treatment1dose,
  HS_1,
  E_inf_1,
  EC50_1,
  treatment2dose,
  HS_2,
  E_inf_2,
  EC50_2,
  tol = 0.1,
  lower_bound = 0,
  upper_bound = 1,
  verbose = FALSE
)

Arguments

treatment1dose

numeric a vector of concentrations for treatment 1

HS_1

numeric Hill coefficient of treatment 1

E_inf_1

numeric viability produced by the maximum attainable effect of treatment 1.

EC50_1

numeric relative EC50 of treatment 1.

treatment2dose

numeric a vector of concentrations for treatment 2

HS_2

numeric Hill coefficient of treatment 2

E_inf_2

numeric viability produced by the maximum attainable effect of treatment 2.

EC50_2

numeric relative EC50 of treatment 2.

tol

numeric Error tolerance for deviations from Loewe assumption. Loewe predictions with error higher than tol will be returned as NA. Deafult 0.1.

lower_bound

numeric Lowest possible value for Loewe expected viability. Default 0.

upper_bound

numeric Highest possible value for Loewe expected viability. Default 1.

verbose

logical whether to display warning messages. Default FALSE.

Value

numeric expected viability under Loewe additive null assumption.

Examples

if (FALSE) {
tre |>
    endoaggregate(
        assay="combo_viability",
        Loewe = computeLoewe(
            treatment1dose=treatment1dose,
            treatment2dose=treatment2dose,
            HS_1=HS_1,
            HS_2=HS_2,
            E_inf_1=E_inf_1,
            E_inf_2=E_inf_2,
            EC50_1=EC50_1,
            EC50_2=EC50_2
        ),
        by = assayKeys(tre, "combo_viability")
    ) -> tre
}