Compute ZIP delta score

Following the calculation of ZIP delta score as in Appendix A3. See reference for details.

Compute ZIP delta score as described in the original paper.

Usage

# S4 method for TreatmentResponseExperiment
computeZIPdelta(object, residual = "logcosh", nthread = 1L, show_Rsqr = FALSE)

Arguments

object

TreatmentResponseExperiment The TreatmentResponseExperiment from which to extract assays mono_profile and combo_viability to compute ZIP delta scores.

residual

character Method used to minimise residual in fitting curves. 3 methods available: c("logcosh", "normal", "Cauchy"). The default method is logcosh. It minimises the logarithmic hyperbolic cosine loss of the residuals and provides the fastest estimation among the three methods, with fitting quality in between normal and Cauchy; recommanded when fitting large-scale datasets. The other two methods minimise residuals by considering the truncated probability distribution (as in their names) for the residual. Cauchy provides the best fitting quality but also takes the longest to run.

nthread

integer Number of cores used to perform computation. Default 1.

show_Rsqr

logical Whether to show the 2-way curve fitting quality in the result. Default FALSE.

Value

TreatmentResponseExperiment with assay combo_scores containing delta_scores

References

Yadav, B., Wennerberg, K., Aittokallio, T., & Tang, J. (2015). Searching for Drug Synergy in Complex Dose–Response Landscapes Using an Interaction Potency Model. Computational and Structural Biotechnology Journal, 13, 504–513. https://doi.org/10.1016/j.csbj.2015.09.001

Examples

if (FALSE) {
tre <- computeZIPdelta(tre, residual = "Cauchy", nthread = 2L)
}