Download Drug Perturbation Signatures

This function allows you to download an array of drug perturbation signatures, as would be computed by the drugPerturbationSig function, for the available perturbation PharmacoSets. This function allows the user to skip these very lengthy calculation steps for the datasets available, and start their analysis from the already computed signatures

Usage

downloadPertSig(
  name,
  saveDir = file.path(".", "PSets", "Sigs"),
  fileName,
  verbose = TRUE,
  ...,
  myfn
)

Arguments

name

A character(1) string, the name of the PharmacoSet for which to download signatures. The name should match the names returned in the PSet Name column of availablePSets(canonical=FALSE).

saveDir

A character(1) string with the folder path where the PharmacoSet should be saved. Defaults to "./PSets/Sigs/". Will create directory if it does not exist.

fileName

character(1) What to name the downloaded file. Defaults to 'name_signature.RData' when excluded.

verbose

logical(1) Should downloader show detailed messages?

...

pairlist Force subsequent arguments to be named.

myfn

character(1) A deprecated version of fileName. Still works for now, but will be deprecated in future releases.

Value

An array type object contaning the signatures

Examples

if (FALSE) {
    if (interactive()) downloadPertSig("CMAP_2016")
}