Viability measurements in dose-reponse curves must remain stable or decrease monotonically reflecting response to the drug being tested. filterNoisyCurves flags dose-response curves that strongly violate these assumptions.

Viability measurements in dose-reponse curves must remain stable or decrease monotonically reflecting response to the drug being tested. filterNoisyCurves flags dose-response curves that strongly violate these assumptions.

Usage

filterNoisyCurves(
  pSet,
  epsilon = 25,
  positive.cutoff.percent = 0.8,
  mean.viablity = 200,
  nthread = 1
)

Arguments

pSet

PharmacoSet a PharmacoSet object

epsilon

numeric a value indicates assumed threshold for the distance between to consecutive viability values on the drug-response curve in the analysis, out of dna, rna, rnaseq, snp, cnv

positive.cutoff.percent

numeric This value indicates that function may violate epsilon rule for how many points on drug-response curve

mean.viablity

numeric average expected viability value

nthread

numeric if multiple cores are available, how many cores should the computation be parallelized over?

Value

a list with two elements 'noisy' containing the rownames of the noisy curves, and 'ok' containing the rownames of the non-noisy curves

Examples

data(GDSCsmall)
filterNoisyCurves(GDSCsmall)
#> $noisy
#>  [1] "drugid_1047_5637"     "drugid_59_8-MG-BA"    "drugid_197_786-0"    
#>  [4] "drugid_1006_786-0"    "drugid_1009_786-0"    "drugid_1012_786-0"   
#>  [7] "drugid_1021_786-0"    "drugid_1022_786-0"    "drugid_1024_786-0"   
#> [10] "drugid_1023_647-V"    "drugid_156_639-V"     "drugid_158_639-V"    
#> [13] "drugid_1014_769-P"    "drugid_1029_769-P"    "drugid_1042_769-P"   
#> [16] "drugid_1050_769-P"    "drugid_1053_769-P"    "drugid_1054_769-P"   
#> [19] "drugid_1057_769-P"    "drugid_1058_769-P"    "drugid_1061_769-P"   
#> [22] "drugid_1062_769-P"    "drugid_1072_769-P"    "drugid_184_23132-87" 
#> [25] "drugid_185_23132-87"  "drugid_186_23132-87"  "drugid_192_23132-87" 
#> [28] "drugid_197_23132-87"  "drugid_1062_23132-87" "drugid_207_22RV1"    
#> [31] "drugid_1021_22RV1"    "drugid_1028_22RV1"   
#> 
#> $ok
#>   [1] "drugid_133_5637"      "drugid_134_5637"      "drugid_135_5637"     
#>   [4] "drugid_136_5637"      "drugid_140_5637"      "drugid_147_5637"     
#>   [7] "drugid_150_5637"      "drugid_151_5637"      "drugid_153_5637"     
#>  [10] "drugid_154_5637"      "drugid_155_5637"      "drugid_156_5637"     
#>  [13] "drugid_157_5637"      "drugid_158_5637"      "drugid_165_5637"     
#>  [16] "drugid_166_5637"      "drugid_167_5637"      "drugid_170_5637"     
#>  [19] "drugid_171_5637"      "drugid_172_5637"      "drugid_173_5637"     
#>  [22] "drugid_175_5637"      "drugid_176_5637"      "drugid_177_5637"     
#>  [25] "drugid_178_5637"      "drugid_180_5637"      "drugid_182_5637"     
#>  [28] "drugid_184_5637"      "drugid_185_5637"      "drugid_186_5637"     
#>  [31] "drugid_190_5637"      "drugid_192_5637"      "drugid_194_5637"     
#>  [34] "drugid_197_5637"      "drugid_199_5637"      "drugid_200_5637"     
#>  [37] "drugid_201_5637"      "drugid_202_5637"      "drugid_204_5637"     
#>  [40] "drugid_207_5637"      "drugid_1001_5637"     "drugid_1003_5637"    
#>  [43] "drugid_1004_5637"     "drugid_1005_5637"     "drugid_1006_5637"    
#>  [46] "drugid_1007_5637"     "drugid_1008_5637"     "drugid_1009_5637"    
#>  [49] "drugid_1010_5637"     "drugid_1011_5637"     "drugid_1012_5637"    
#>  [52] "drugid_1013_5637"     "drugid_1014_5637"     "drugid_1015_5637"    
#>  [55] "drugid_1016_5637"     "drugid_1017_5637"     "drugid_1018_5637"    
#>  [58] "drugid_1019_5637"     "drugid_1020_5637"     "drugid_1021_5637"    
#>  [61] "drugid_1022_5637"     "drugid_1023_5637"     "drugid_1024_5637"    
#>  [64] "drugid_1025_5637"     "drugid_1026_5637"     "drugid_1028_5637"    
#>  [67] "drugid_1029_5637"     "drugid_1030_5637"     "drugid_1031_5637"    
#>  [70] "drugid_1032_5637"     "drugid_1033_5637"     "drugid_1036_5637"    
#>  [73] "drugid_1037_5637"     "drugid_1038_5637"     "drugid_1039_5637"    
#>  [76] "drugid_1042_5637"     "drugid_1043_5637"     "drugid_1046_5637"    
#>  [79] "drugid_1049_5637"     "drugid_1050_5637"     "drugid_1052_5637"    
#>  [82] "drugid_1053_5637"     "drugid_1054_5637"     "drugid_1057_5637"    
#>  [85] "drugid_1058_5637"     "drugid_1059_5637"     "drugid_1060_5637"    
#>  [88] "drugid_1061_5637"     "drugid_1062_5637"     "drugid_1066_5637"    
#>  [91] "drugid_1067_5637"     "drugid_1069_5637"     "drugid_1072_5637"    
#>  [94] "drugid_1129_5637"     "drugid_1133_5637"     "drugid_1149_5637"    
#>  [97] "drugid_1170_5637"     "drugid_1175_5637"     "drugid_1_8-MG-BA"    
#> [100] "drugid_3_8-MG-BA"     "drugid_5_8-MG-BA"     "drugid_6_8-MG-BA"    
#> [103] "drugid_9_8-MG-BA"     "drugid_11_8-MG-BA"    "drugid_17_8-MG-BA"   
#> [106] "drugid_29_8-MG-BA"    "drugid_30_8-MG-BA"    "drugid_32_8-MG-BA"   
#> [109] "drugid_34_8-MG-BA"    "drugid_35_8-MG-BA"    "drugid_37_8-MG-BA"   
#> [112] "drugid_38_8-MG-BA"    "drugid_41_8-MG-BA"    "drugid_45_8-MG-BA"   
#> [115] "drugid_51_8-MG-BA"    "drugid_52_8-MG-BA"    "drugid_53_8-MG-BA"   
#> [118] "drugid_54_8-MG-BA"    "drugid_55_8-MG-BA"    "drugid_56_8-MG-BA"   
#> [121] "drugid_60_8-MG-BA"    "drugid_62_8-MG-BA"    "drugid_63_8-MG-BA"   
#> [124] "drugid_64_8-MG-BA"    "drugid_71_8-MG-BA"    "drugid_83_8-MG-BA"   
#> [127] "drugid_86_8-MG-BA"    "drugid_87_8-MG-BA"    "drugid_88_8-MG-BA"   
#> [130] "drugid_89_8-MG-BA"    "drugid_91_8-MG-BA"    "drugid_94_8-MG-BA"   
#> [133] "drugid_104_8-MG-BA"   "drugid_106_8-MG-BA"   "drugid_110_8-MG-BA"  
#> [136] "drugid_111_8-MG-BA"   "drugid_119_8-MG-BA"   "drugid_127_8-MG-BA"  
#> [139] "drugid_133_8-MG-BA"   "drugid_134_8-MG-BA"   "drugid_135_8-MG-BA"  
#> [142] "drugid_136_8-MG-BA"   "drugid_140_8-MG-BA"   "drugid_147_8-MG-BA"  
#> [145] "drugid_150_8-MG-BA"   "drugid_151_8-MG-BA"   "drugid_153_8-MG-BA"  
#> [148] "drugid_154_8-MG-BA"   "drugid_155_8-MG-BA"   "drugid_156_8-MG-BA"  
#> [151] "drugid_157_8-MG-BA"   "drugid_158_8-MG-BA"   "drugid_165_8-MG-BA"  
#> [154] "drugid_166_8-MG-BA"   "drugid_167_8-MG-BA"   "drugid_170_8-MG-BA"  
#> [157] "drugid_171_8-MG-BA"   "drugid_172_8-MG-BA"   "drugid_173_8-MG-BA"  
#> [160] "drugid_175_8-MG-BA"   "drugid_176_8-MG-BA"   "drugid_177_8-MG-BA"  
#> [163] "drugid_178_8-MG-BA"   "drugid_180_8-MG-BA"   "drugid_182_8-MG-BA"  
#> [166] "drugid_184_8-MG-BA"   "drugid_185_8-MG-BA"   "drugid_186_8-MG-BA"  
#> [169] "drugid_190_8-MG-BA"   "drugid_192_8-MG-BA"   "drugid_194_8-MG-BA"  
#> [172] "drugid_197_8-MG-BA"   "drugid_199_8-MG-BA"   "drugid_200_8-MG-BA"  
#> [175] "drugid_201_8-MG-BA"   "drugid_202_8-MG-BA"   "drugid_204_8-MG-BA"  
#> [178] "drugid_207_8-MG-BA"   "drugid_1001_8-MG-BA"  "drugid_1003_8-MG-BA" 
#> [181] "drugid_1004_8-MG-BA"  "drugid_1005_8-MG-BA"  "drugid_1006_8-MG-BA" 
#> [184] "drugid_1007_8-MG-BA"  "drugid_1008_8-MG-BA"  "drugid_1009_8-MG-BA" 
#> [187] "drugid_1010_8-MG-BA"  "drugid_1011_8-MG-BA"  "drugid_1012_8-MG-BA" 
#> [190] "drugid_1013_8-MG-BA"  "drugid_1014_8-MG-BA"  "drugid_1015_8-MG-BA" 
#> [193] "drugid_1016_8-MG-BA"  "drugid_1017_8-MG-BA"  "drugid_1018_8-MG-BA" 
#> [196] "drugid_1019_8-MG-BA"  "drugid_1020_8-MG-BA"  "drugid_1021_8-MG-BA" 
#> [199] "drugid_1022_8-MG-BA"  "drugid_1023_8-MG-BA"  "drugid_1024_8-MG-BA" 
#> [202] "drugid_1025_8-MG-BA"  "drugid_1026_8-MG-BA"  "drugid_1028_8-MG-BA" 
#> [205] "drugid_1029_8-MG-BA"  "drugid_1030_8-MG-BA"  "drugid_1031_8-MG-BA" 
#> [208] "drugid_1032_8-MG-BA"  "drugid_1033_8-MG-BA"  "drugid_1036_8-MG-BA" 
#> [211] "drugid_1037_8-MG-BA"  "drugid_1038_8-MG-BA"  "drugid_1039_8-MG-BA" 
#> [214] "drugid_1042_8-MG-BA"  "drugid_1043_8-MG-BA"  "drugid_1046_8-MG-BA" 
#> [217] "drugid_1047_8-MG-BA"  "drugid_1049_8-MG-BA"  "drugid_1050_8-MG-BA" 
#> [220] "drugid_1052_8-MG-BA"  "drugid_1053_8-MG-BA"  "drugid_1054_8-MG-BA" 
#> [223] "drugid_1057_8-MG-BA"  "drugid_1058_8-MG-BA"  "drugid_1059_8-MG-BA" 
#> [226] "drugid_1060_8-MG-BA"  "drugid_1061_8-MG-BA"  "drugid_1062_8-MG-BA" 
#> [229] "drugid_1066_8-MG-BA"  "drugid_1067_8-MG-BA"  "drugid_1069_8-MG-BA" 
#> [232] "drugid_1072_8-MG-BA"  "drugid_1129_8-MG-BA"  "drugid_1133_8-MG-BA" 
#> [235] "drugid_1149_8-MG-BA"  "drugid_1170_8-MG-BA"  "drugid_1175_8-MG-BA" 
#> [238] "drugid_133_786-0"     "drugid_134_786-0"     "drugid_135_786-0"    
#> [241] "drugid_136_786-0"     "drugid_140_786-0"     "drugid_147_786-0"    
#> [244] "drugid_150_786-0"     "drugid_151_786-0"     "drugid_153_786-0"    
#> [247] "drugid_154_786-0"     "drugid_155_786-0"     "drugid_156_786-0"    
#> [250] "drugid_157_786-0"     "drugid_158_786-0"     "drugid_165_786-0"    
#> [253] "drugid_166_786-0"     "drugid_167_786-0"     "drugid_170_786-0"    
#> [256] "drugid_171_786-0"     "drugid_172_786-0"     "drugid_173_786-0"    
#> [259] "drugid_175_786-0"     "drugid_176_786-0"     "drugid_177_786-0"    
#> [262] "drugid_178_786-0"     "drugid_180_786-0"     "drugid_182_786-0"    
#> [265] "drugid_184_786-0"     "drugid_185_786-0"     "drugid_186_786-0"    
#> [268] "drugid_190_786-0"     "drugid_192_786-0"     "drugid_194_786-0"    
#> [271] "drugid_199_786-0"     "drugid_200_786-0"     "drugid_201_786-0"    
#> [274] "drugid_202_786-0"     "drugid_204_786-0"     "drugid_207_786-0"    
#> [277] "drugid_1001_786-0"    "drugid_1003_786-0"    "drugid_1004_786-0"   
#> [280] "drugid_1005_786-0"    "drugid_1007_786-0"    "drugid_1008_786-0"   
#> [283] "drugid_1010_786-0"    "drugid_1011_786-0"    "drugid_1013_786-0"   
#> [286] "drugid_1014_786-0"    "drugid_1015_786-0"    "drugid_1016_786-0"   
#> [289] "drugid_1017_786-0"    "drugid_1018_786-0"    "drugid_1019_786-0"   
#> [292] "drugid_1020_786-0"    "drugid_1023_786-0"    "drugid_1025_786-0"   
#> [295] "drugid_1026_786-0"    "drugid_1028_786-0"    "drugid_1029_786-0"   
#> [298] "drugid_1030_786-0"    "drugid_1031_786-0"    "drugid_1032_786-0"   
#> [301] "drugid_1033_786-0"    "drugid_1036_786-0"    "drugid_1037_786-0"   
#> [304] "drugid_1038_786-0"    "drugid_1039_786-0"    "drugid_1042_786-0"   
#> [307] "drugid_1043_786-0"    "drugid_1046_786-0"    "drugid_1047_786-0"   
#> [310] "drugid_1049_786-0"    "drugid_1050_786-0"    "drugid_1052_786-0"   
#> [313] "drugid_1053_786-0"    "drugid_1054_786-0"    "drugid_1057_786-0"   
#> [316] "drugid_1058_786-0"    "drugid_1059_786-0"    "drugid_1060_786-0"   
#> [319] "drugid_1061_786-0"    "drugid_1062_786-0"    "drugid_1066_786-0"   
#> [322] "drugid_1067_786-0"    "drugid_1069_786-0"    "drugid_1072_786-0"   
#> [325] "drugid_1129_786-0"    "drugid_1133_786-0"    "drugid_1149_786-0"   
#> [328] "drugid_1170_786-0"    "drugid_1175_786-0"    "drugid_133_647-V"    
#> [331] "drugid_134_647-V"     "drugid_135_647-V"     "drugid_136_647-V"    
#> [334] "drugid_140_647-V"     "drugid_147_647-V"     "drugid_150_647-V"    
#> [337] "drugid_151_647-V"     "drugid_153_647-V"     "drugid_154_647-V"    
#> [340] "drugid_155_647-V"     "drugid_156_647-V"     "drugid_157_647-V"    
#> [343] "drugid_158_647-V"     "drugid_165_647-V"     "drugid_166_647-V"    
#> [346] "drugid_167_647-V"     "drugid_170_647-V"     "drugid_171_647-V"    
#> [349] "drugid_172_647-V"     "drugid_173_647-V"     "drugid_175_647-V"    
#> [352] "drugid_176_647-V"     "drugid_177_647-V"     "drugid_178_647-V"    
#> [355] "drugid_180_647-V"     "drugid_182_647-V"     "drugid_184_647-V"    
#> [358] "drugid_185_647-V"     "drugid_186_647-V"     "drugid_190_647-V"    
#> [361] "drugid_192_647-V"     "drugid_194_647-V"     "drugid_197_647-V"    
#> [364] "drugid_199_647-V"     "drugid_200_647-V"     "drugid_201_647-V"    
#> [367] "drugid_202_647-V"     "drugid_204_647-V"     "drugid_207_647-V"    
#> [370] "drugid_1001_647-V"    "drugid_1003_647-V"    "drugid_1004_647-V"   
#> [373] "drugid_1005_647-V"    "drugid_1006_647-V"    "drugid_1007_647-V"   
#> [376] "drugid_1008_647-V"    "drugid_1009_647-V"    "drugid_1010_647-V"   
#> [379] "drugid_1011_647-V"    "drugid_1012_647-V"    "drugid_1013_647-V"   
#> [382] "drugid_1014_647-V"    "drugid_1015_647-V"    "drugid_1016_647-V"   
#> [385] "drugid_1017_647-V"    "drugid_1018_647-V"    "drugid_1019_647-V"   
#> [388] "drugid_1020_647-V"    "drugid_1021_647-V"    "drugid_1022_647-V"   
#> [391] "drugid_1024_647-V"    "drugid_1025_647-V"    "drugid_1026_647-V"   
#> [394] "drugid_1028_647-V"    "drugid_1029_647-V"    "drugid_1030_647-V"   
#> [397] "drugid_1031_647-V"    "drugid_1032_647-V"    "drugid_1033_647-V"   
#> [400] "drugid_1036_647-V"    "drugid_1037_647-V"    "drugid_1038_647-V"   
#> [403] "drugid_1039_647-V"    "drugid_1042_647-V"    "drugid_1043_647-V"   
#> [406] "drugid_1046_647-V"    "drugid_1047_647-V"    "drugid_1049_647-V"   
#> [409] "drugid_1050_647-V"    "drugid_1052_647-V"    "drugid_1053_647-V"   
#> [412] "drugid_1054_647-V"    "drugid_1057_647-V"    "drugid_1058_647-V"   
#> [415] "drugid_1059_647-V"    "drugid_1060_647-V"    "drugid_1061_647-V"   
#> [418] "drugid_1062_647-V"    "drugid_1066_647-V"    "drugid_1067_647-V"   
#> [421] "drugid_1069_647-V"    "drugid_1072_647-V"    "drugid_1129_647-V"   
#> [424] "drugid_1133_647-V"    "drugid_1149_647-V"    "drugid_1170_647-V"   
#> [427] "drugid_1175_647-V"    "drugid_133_639-V"     "drugid_134_639-V"    
#> [430] "drugid_135_639-V"     "drugid_136_639-V"     "drugid_140_639-V"    
#> [433] "drugid_147_639-V"     "drugid_150_639-V"     "drugid_151_639-V"    
#> [436] "drugid_153_639-V"     "drugid_154_639-V"     "drugid_155_639-V"    
#> [439] "drugid_157_639-V"     "drugid_165_639-V"     "drugid_166_639-V"    
#> [442] "drugid_167_639-V"     "drugid_170_639-V"     "drugid_171_639-V"    
#> [445] "drugid_172_639-V"     "drugid_173_639-V"     "drugid_175_639-V"    
#> [448] "drugid_176_639-V"     "drugid_177_639-V"     "drugid_178_639-V"    
#> [451] "drugid_180_639-V"     "drugid_182_639-V"     "drugid_184_639-V"    
#> [454] "drugid_185_639-V"     "drugid_186_639-V"     "drugid_190_639-V"    
#> [457] "drugid_192_639-V"     "drugid_194_639-V"     "drugid_197_639-V"    
#> [460] "drugid_199_639-V"     "drugid_200_639-V"     "drugid_201_639-V"    
#> [463] "drugid_202_639-V"     "drugid_204_639-V"     "drugid_207_639-V"    
#> [466] "drugid_1001_639-V"    "drugid_1003_639-V"    "drugid_1004_639-V"   
#> [469] "drugid_1005_639-V"    "drugid_1006_639-V"    "drugid_1007_639-V"   
#> [472] "drugid_1008_639-V"    "drugid_1009_639-V"    "drugid_1010_639-V"   
#> [475] "drugid_1011_639-V"    "drugid_1012_639-V"    "drugid_1013_639-V"   
#> [478] "drugid_1014_639-V"    "drugid_1015_639-V"    "drugid_1016_639-V"   
#> [481] "drugid_1017_639-V"    "drugid_1018_639-V"    "drugid_1019_639-V"   
#> [484] "drugid_1020_639-V"    "drugid_1021_639-V"    "drugid_1022_639-V"   
#> [487] "drugid_1023_639-V"    "drugid_1024_639-V"    "drugid_1025_639-V"   
#> [490] "drugid_1026_639-V"    "drugid_1028_639-V"    "drugid_1029_639-V"   
#> [493] "drugid_1030_639-V"    "drugid_1031_639-V"    "drugid_1032_639-V"   
#> [496] "drugid_1033_639-V"    "drugid_1036_639-V"    "drugid_1037_639-V"   
#> [499] "drugid_1038_639-V"    "drugid_1039_639-V"    "drugid_1042_639-V"   
#> [502] "drugid_1043_639-V"    "drugid_1046_639-V"    "drugid_1047_639-V"   
#> [505] "drugid_1049_639-V"    "drugid_1050_639-V"    "drugid_1052_639-V"   
#> [508] "drugid_1053_639-V"    "drugid_1054_639-V"    "drugid_1057_639-V"   
#> [511] "drugid_1058_639-V"    "drugid_1059_639-V"    "drugid_1060_639-V"   
#> [514] "drugid_1061_639-V"    "drugid_1062_639-V"    "drugid_1066_639-V"   
#> [517] "drugid_1067_639-V"    "drugid_1069_639-V"    "drugid_1072_639-V"   
#> [520] "drugid_1129_639-V"    "drugid_1133_639-V"    "drugid_1149_639-V"   
#> [523] "drugid_1170_639-V"    "drugid_1175_639-V"    "drugid_1_697"        
#> [526] "drugid_3_697"         "drugid_5_697"         "drugid_6_697"        
#> [529] "drugid_9_697"         "drugid_11_697"        "drugid_17_697"       
#> [532] "drugid_29_697"        "drugid_30_697"        "drugid_32_697"       
#> [535] "drugid_34_697"        "drugid_35_697"        "drugid_37_697"       
#> [538] "drugid_38_697"        "drugid_41_697"        "drugid_45_697"       
#> [541] "drugid_51_697"        "drugid_52_697"        "drugid_53_697"       
#> [544] "drugid_54_697"        "drugid_55_697"        "drugid_56_697"       
#> [547] "drugid_59_697"        "drugid_60_697"        "drugid_62_697"       
#> [550] "drugid_63_697"        "drugid_64_697"        "drugid_71_697"       
#> [553] "drugid_83_697"        "drugid_86_697"        "drugid_87_697"       
#> [556] "drugid_88_697"        "drugid_89_697"        "drugid_91_697"       
#> [559] "drugid_94_697"        "drugid_104_697"       "drugid_106_697"      
#> [562] "drugid_110_697"       "drugid_111_697"       "drugid_119_697"      
#> [565] "drugid_127_697"       "drugid_133_697"       "drugid_134_697"      
#> [568] "drugid_135_697"       "drugid_136_697"       "drugid_140_697"      
#> [571] "drugid_147_697"       "drugid_150_697"       "drugid_151_697"      
#> [574] "drugid_153_697"       "drugid_154_697"       "drugid_155_697"      
#> [577] "drugid_156_697"       "drugid_157_697"       "drugid_158_697"      
#> [580] "drugid_165_697"       "drugid_166_697"       "drugid_167_697"      
#> [583] "drugid_170_697"       "drugid_171_697"       "drugid_172_697"      
#> [586] "drugid_173_697"       "drugid_175_697"       "drugid_176_697"      
#> [589] "drugid_177_697"       "drugid_178_697"       "drugid_180_697"      
#> [592] "drugid_182_697"       "drugid_184_697"       "drugid_185_697"      
#> [595] "drugid_186_697"       "drugid_190_697"       "drugid_192_697"      
#> [598] "drugid_194_697"       "drugid_197_697"       "drugid_199_697"      
#> [601] "drugid_200_697"       "drugid_201_697"       "drugid_202_697"      
#> [604] "drugid_204_697"       "drugid_207_697"       "drugid_1001_697"     
#> [607] "drugid_1003_697"      "drugid_1004_697"      "drugid_1005_697"     
#> [610] "drugid_1006_697"      "drugid_1007_697"      "drugid_1008_697"     
#> [613] "drugid_1009_697"      "drugid_1010_697"      "drugid_1011_697"     
#> [616] "drugid_1012_697"      "drugid_1013_697"      "drugid_1014_697"     
#> [619] "drugid_1015_697"      "drugid_1016_697"      "drugid_1017_697"     
#> [622] "drugid_1018_697"      "drugid_1019_697"      "drugid_1020_697"     
#> [625] "drugid_1021_697"      "drugid_1022_697"      "drugid_1023_697"     
#> [628] "drugid_1024_697"      "drugid_1025_697"      "drugid_1026_697"     
#> [631] "drugid_1028_697"      "drugid_1029_697"      "drugid_1030_697"     
#> [634] "drugid_1031_697"      "drugid_1032_697"      "drugid_1033_697"     
#> [637] "drugid_1036_697"      "drugid_1037_697"      "drugid_1038_697"     
#> [640] "drugid_1039_697"      "drugid_1042_697"      "drugid_1043_697"     
#> [643] "drugid_1046_697"      "drugid_1047_697"      "drugid_1049_697"     
#> [646] "drugid_1050_697"      "drugid_1052_697"      "drugid_1053_697"     
#> [649] "drugid_1054_697"      "drugid_1057_697"      "drugid_1058_697"     
#> [652] "drugid_1059_697"      "drugid_1060_697"      "drugid_1061_697"     
#> [655] "drugid_1062_697"      "drugid_1066_697"      "drugid_1067_697"     
#> [658] "drugid_1069_697"      "drugid_1072_697"      "drugid_1129_697"     
#> [661] "drugid_1133_697"      "drugid_1149_697"      "drugid_1170_697"     
#> [664] "drugid_1175_697"      "drugid_133_769-P"     "drugid_134_769-P"    
#> [667] "drugid_135_769-P"     "drugid_136_769-P"     "drugid_140_769-P"    
#> [670] "drugid_147_769-P"     "drugid_150_769-P"     "drugid_151_769-P"    
#> [673] "drugid_153_769-P"     "drugid_154_769-P"     "drugid_155_769-P"    
#> [676] "drugid_156_769-P"     "drugid_157_769-P"     "drugid_158_769-P"    
#> [679] "drugid_165_769-P"     "drugid_166_769-P"     "drugid_167_769-P"    
#> [682] "drugid_170_769-P"     "drugid_171_769-P"     "drugid_172_769-P"    
#> [685] "drugid_173_769-P"     "drugid_175_769-P"     "drugid_176_769-P"    
#> [688] "drugid_177_769-P"     "drugid_178_769-P"     "drugid_180_769-P"    
#> [691] "drugid_182_769-P"     "drugid_184_769-P"     "drugid_185_769-P"    
#> [694] "drugid_186_769-P"     "drugid_190_769-P"     "drugid_192_769-P"    
#> [697] "drugid_194_769-P"     "drugid_197_769-P"     "drugid_199_769-P"    
#> [700] "drugid_200_769-P"     "drugid_201_769-P"     "drugid_202_769-P"    
#> [703] "drugid_204_769-P"     "drugid_207_769-P"     "drugid_1001_769-P"   
#> [706] "drugid_1003_769-P"    "drugid_1004_769-P"    "drugid_1005_769-P"   
#> [709] "drugid_1006_769-P"    "drugid_1007_769-P"    "drugid_1008_769-P"   
#> [712] "drugid_1009_769-P"    "drugid_1010_769-P"    "drugid_1011_769-P"   
#> [715] "drugid_1012_769-P"    "drugid_1013_769-P"    "drugid_1015_769-P"   
#> [718] "drugid_1016_769-P"    "drugid_1017_769-P"    "drugid_1018_769-P"   
#> [721] "drugid_1019_769-P"    "drugid_1020_769-P"    "drugid_1021_769-P"   
#> [724] "drugid_1022_769-P"    "drugid_1023_769-P"    "drugid_1024_769-P"   
#> [727] "drugid_1025_769-P"    "drugid_1026_769-P"    "drugid_1028_769-P"   
#> [730] "drugid_1030_769-P"    "drugid_1031_769-P"    "drugid_1032_769-P"   
#> [733] "drugid_1033_769-P"    "drugid_1036_769-P"    "drugid_1037_769-P"   
#> [736] "drugid_1038_769-P"    "drugid_1039_769-P"    "drugid_1043_769-P"   
#> [739] "drugid_1046_769-P"    "drugid_1047_769-P"    "drugid_1049_769-P"   
#> [742] "drugid_1052_769-P"    "drugid_1059_769-P"    "drugid_1060_769-P"   
#> [745] "drugid_1066_769-P"    "drugid_1067_769-P"    "drugid_1069_769-P"   
#> [748] "drugid_1129_769-P"    "drugid_1133_769-P"    "drugid_1149_769-P"   
#> [751] "drugid_1170_769-P"    "drugid_1175_769-P"    "drugid_133_23132-87" 
#> [754] "drugid_134_23132-87"  "drugid_135_23132-87"  "drugid_136_23132-87" 
#> [757] "drugid_140_23132-87"  "drugid_147_23132-87"  "drugid_150_23132-87" 
#> [760] "drugid_151_23132-87"  "drugid_153_23132-87"  "drugid_154_23132-87" 
#> [763] "drugid_155_23132-87"  "drugid_156_23132-87"  "drugid_157_23132-87" 
#> [766] "drugid_158_23132-87"  "drugid_165_23132-87"  "drugid_166_23132-87" 
#> [769] "drugid_167_23132-87"  "drugid_170_23132-87"  "drugid_171_23132-87" 
#> [772] "drugid_172_23132-87"  "drugid_173_23132-87"  "drugid_175_23132-87" 
#> [775] "drugid_176_23132-87"  "drugid_177_23132-87"  "drugid_178_23132-87" 
#> [778] "drugid_180_23132-87"  "drugid_182_23132-87"  "drugid_190_23132-87" 
#> [781] "drugid_194_23132-87"  "drugid_199_23132-87"  "drugid_200_23132-87" 
#> [784] "drugid_201_23132-87"  "drugid_202_23132-87"  "drugid_204_23132-87" 
#> [787] "drugid_207_23132-87"  "drugid_1001_23132-87" "drugid_1003_23132-87"
#> [790] "drugid_1004_23132-87" "drugid_1005_23132-87" "drugid_1006_23132-87"
#> [793] "drugid_1007_23132-87" "drugid_1008_23132-87" "drugid_1009_23132-87"
#> [796] "drugid_1010_23132-87" "drugid_1011_23132-87" "drugid_1012_23132-87"
#> [799] "drugid_1013_23132-87" "drugid_1014_23132-87" "drugid_1015_23132-87"
#> [802] "drugid_1016_23132-87" "drugid_1017_23132-87" "drugid_1018_23132-87"
#> [805] "drugid_1019_23132-87" "drugid_1020_23132-87" "drugid_1021_23132-87"
#> [808] "drugid_1022_23132-87" "drugid_1023_23132-87" "drugid_1024_23132-87"
#> [811] "drugid_1026_23132-87" "drugid_1028_23132-87" "drugid_1029_23132-87"
#> [814] "drugid_1030_23132-87" "drugid_1031_23132-87" "drugid_1032_23132-87"
#> [817] "drugid_1033_23132-87" "drugid_1036_23132-87" "drugid_1037_23132-87"
#> [820] "drugid_1038_23132-87" "drugid_1039_23132-87" "drugid_1042_23132-87"
#> [823] "drugid_1043_23132-87" "drugid_1046_23132-87" "drugid_1047_23132-87"
#> [826] "drugid_1049_23132-87" "drugid_1050_23132-87" "drugid_1052_23132-87"
#> [829] "drugid_1053_23132-87" "drugid_1054_23132-87" "drugid_1057_23132-87"
#> [832] "drugid_1058_23132-87" "drugid_1059_23132-87" "drugid_1060_23132-87"
#> [835] "drugid_1061_23132-87" "drugid_1066_23132-87" "drugid_1067_23132-87"
#> [838] "drugid_1069_23132-87" "drugid_1072_23132-87" "drugid_1129_23132-87"
#> [841] "drugid_1133_23132-87" "drugid_1149_23132-87" "drugid_1170_23132-87"
#> [844] "drugid_1175_23132-87" "drugid_133_22RV1"     "drugid_134_22RV1"    
#> [847] "drugid_135_22RV1"     "drugid_136_22RV1"     "drugid_140_22RV1"    
#> [850] "drugid_147_22RV1"     "drugid_150_22RV1"     "drugid_151_22RV1"    
#> [853] "drugid_153_22RV1"     "drugid_154_22RV1"     "drugid_155_22RV1"    
#> [856] "drugid_156_22RV1"     "drugid_157_22RV1"     "drugid_158_22RV1"    
#> [859] "drugid_165_22RV1"     "drugid_166_22RV1"     "drugid_167_22RV1"    
#> [862] "drugid_170_22RV1"     "drugid_171_22RV1"     "drugid_172_22RV1"    
#> [865] "drugid_173_22RV1"     "drugid_175_22RV1"     "drugid_176_22RV1"    
#> [868] "drugid_177_22RV1"     "drugid_178_22RV1"     "drugid_180_22RV1"    
#> [871] "drugid_182_22RV1"     "drugid_184_22RV1"     "drugid_185_22RV1"    
#> [874] "drugid_186_22RV1"     "drugid_190_22RV1"     "drugid_192_22RV1"    
#> [877] "drugid_194_22RV1"     "drugid_197_22RV1"     "drugid_199_22RV1"    
#> [880] "drugid_200_22RV1"     "drugid_201_22RV1"     "drugid_202_22RV1"    
#> [883] "drugid_204_22RV1"     "drugid_1001_22RV1"    "drugid_1003_22RV1"   
#> [886] "drugid_1004_22RV1"    "drugid_1005_22RV1"    "drugid_1006_22RV1"   
#> [889] "drugid_1007_22RV1"    "drugid_1008_22RV1"    "drugid_1009_22RV1"   
#> [892] "drugid_1010_22RV1"    "drugid_1011_22RV1"    "drugid_1012_22RV1"   
#> [895] "drugid_1013_22RV1"    "drugid_1014_22RV1"    "drugid_1015_22RV1"   
#> [898] "drugid_1016_22RV1"    "drugid_1017_22RV1"    "drugid_1018_22RV1"   
#> [901] "drugid_1019_22RV1"    "drugid_1020_22RV1"    "drugid_1022_22RV1"   
#> [904] "drugid_1023_22RV1"    "drugid_1024_22RV1"    "drugid_1025_22RV1"   
#> [907] "drugid_1026_22RV1"    "drugid_1029_22RV1"    "drugid_1030_22RV1"   
#> [910] "drugid_1031_22RV1"    "drugid_1032_22RV1"    "drugid_1033_22RV1"   
#> [913] "drugid_1036_22RV1"    "drugid_1037_22RV1"    "drugid_1038_22RV1"   
#> [916] "drugid_1039_22RV1"    "drugid_1042_22RV1"    "drugid_1043_22RV1"   
#> [919] "drugid_1046_22RV1"    "drugid_1047_22RV1"    "drugid_1049_22RV1"   
#> [922] "drugid_1050_22RV1"    "drugid_1052_22RV1"    "drugid_1053_22RV1"   
#> [925] "drugid_1054_22RV1"    "drugid_1057_22RV1"    "drugid_1058_22RV1"   
#> [928] "drugid_1059_22RV1"    "drugid_1060_22RV1"    "drugid_1061_22RV1"   
#> [931] "drugid_1062_22RV1"    "drugid_1066_22RV1"    "drugid_1067_22RV1"   
#> [934] "drugid_1069_22RV1"    "drugid_1072_22RV1"    "drugid_1129_22RV1"   
#> [937] "drugid_1133_22RV1"    "drugid_1149_22RV1"    "drugid_1170_22RV1"   
#> [940] "drugid_1175_22RV1"   
#>