A function to subset a PharmacoSet to data containing only specified drugs, cells and genes

This is the prefered method of subsetting a PharmacoSet. This function allows abstraction of the data to the level of biologically relevant objects: drugs and cells. The function will automatically go through all of the combined data in the PharmacoSet and ensure only the requested drugs and cell lines are found in any of the slots. This allows quickly picking out all the experiments for a drug or cell of interest, as well removes the need to keep track of all the metadata conventions between different datasets.

Usage

# S4 method for PharmacoSet
subsetTo(
  object,
  cells = NULL,
  drugs = NULL,
  molecular.data.cells = NULL,
  keep.controls = TRUE,
  ...
)

Arguments

object

A PharmacoSet to be subsetted

cells

A list or vector of cell names as used in the dataset to which the object will be subsetted. If left blank, then all cells will be left in the dataset.

drugs

A list or vector of drug names as used in the dataset to which the object will be subsetted. If left blank, then all drugs will be left in the dataset.

molecular.data.cells

A list or vector of cell names to keep in the molecular data

keep.controls

If the dataset has perturbation type experiments, should the controls be kept in the dataset? Defaults to true.

...

Other arguments passed by other function within the package

Value

A PharmacoSet with only the selected drugs and cells

Examples

data(CCLEsmall)
CCLEdrugs  <- treatmentNames(CCLEsmall)
CCLEcells <- sampleNames(CCLEsmall)
pSet <- subsetTo(CCLEsmall, drugs = CCLEdrugs[1], cells = CCLEcells[1])
pSet
#> <PharmacoSet>
#> Name:  CCLE 
#> Date Created:  Fri Nov  6 14:00:53 2015 
#> Number of samples:  1 
#> Molecular profiles:
#> RNA :
#>   Dim: 50, 1 
#> RNAseq :
#>   Dim: 50, 0 
#> mutation :
#>   Dim: 1667, 1 
#> CNV :
#>   Dim: 50, 1 
#> Treatment response:
#> Drug pertubation:
#>    Please look at pertNumber(cSet) to determine number of experiments  for each drug-sample combination.
#> Drug sensitivity:
#>    Number of Experiments:  1 
#>    Please look at sensNumber(cSet) to determine number of  experiments for each drug-sample combination.