Takes the sensitivity data from a PharmacoSet, and summarises them into a drug vs cell line table

This function creates a table with cell lines as rows and drugs as columns, summarising the drug senstitivity data of a PharmacoSet into drug-cell line pairs

Usage

# S4 method for PharmacoSet
summarizeSensitivityProfiles(
  object,
  sensitivity.measure = "auc_recomputed",
  cell.lines,
  drugs,
  summary.stat = c("mean", "median", "first", "last", "max", "min"),
  fill.missing = TRUE,
  verbose = TRUE
)

Arguments

object

PharmacoSet The PharmacoSet from which to extract the data

sensitivity.measure

character which sensitivity sensitivity.measure to use? Use the sensitivityMeasures function to find out what measures are available for each object.

cell.lines

character The cell lines to be summarized. If any cell lines has no data, it will be filled with missing values

drugs

character The drugs to be summarized. If any drugs has no data, it will be filled with missing values

summary.stat

character which summary method to use if there are repeated cell line-drug experiments? Choices are "mean", "median", "first", or "last"

fill.missing

boolean should the missing cell lines not in the molecular data object be filled in with missing values?

verbose

Should the function print progress messages?

Value

matrix A matrix with cell lines going down the rows, drugs across the columns, with the selected sensitivity statistic for each pair.

Examples

data(GDSCsmall)
GDSCauc <- summarizeSensitivityProfiles(GDSCsmall,
    sensitivity.measure='auc_published')