Show the Annotations of a signature object

This funtion prints out the information about the call used to compute the drug signatures, and the session info for the session in which the computation was done. Useful for determining the exact conditions used to generate signatures.

Usage

# S4 method for PharmacoSig
showSigAnnot(object)

Arguments

object

An object of the PharmacoSig Class, as returned by drugPerturbationSig or drugSensitivitySig

Value

Prints the PharmacoGx Signatures annotations to the output stream, and returns invisible NULL.

Examples

data(GDSCsmall)
drug.sensitivity <- drugSensitivitySig(GDSCsmall, mDataType="rna",
             nthread=1, features = fNames(GDSCsmall, "rna")[1])
#> Summarizing rna molecular data for:	GDSC
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Computing drug sensitivity signatures...
showSigAnnot(drug.sensitivity)
#>  [1] ".drugSensitivitySigPharmacoSet" "object"                        
#>  [3] "mDataType"                      "drugs"                         
#>  [5] "features"                       "cells"                         
#>  [7] "tissues"                        "sensitivity.measure"           
#>  [9] "molecular.summary.stat"         "sensitivity.summary.stat"      
#> [11] "returnValues"                   "sensitivity.cutoff"            
#> [13] "standardize"                    "molecular.cutoff"              
#> [15] "molecular.cutoff.direction"     "nthread"                       
#> [17] "parallel.on"                    "modeling.method"               
#> [19] "inference.method"               "verbose"                       
#> R version 4.2.1 (2022-06-23)
#> Platform: x86_64-pc-linux-gnu (64-bit)
#> Running under: Ubuntu 22.04.2 LTS
#> 
#> Matrix products: default
#> BLAS:   /usr/lib/x86_64-linux-gnu/openblas-pthread/libblas.so.3
#> LAPACK: /usr/lib/x86_64-linux-gnu/openblas-pthread/libopenblasp-r0.3.20.so
#> 
#> locale:
#>  [1] LC_CTYPE=C.UTF-8       LC_NUMERIC=C           LC_TIME=C.UTF-8       
#>  [4] LC_COLLATE=C           LC_MONETARY=C.UTF-8    LC_MESSAGES=C.UTF-8   
#>  [7] LC_PAPER=C.UTF-8       LC_NAME=C              LC_ADDRESS=C          
#> [10] LC_TELEPHONE=C         LC_MEASUREMENT=C.UTF-8 LC_IDENTIFICATION=C   
#> 
#> attached base packages:
#> [1] stats4    stats     graphics  grDevices utils     datasets  methods  
#> [8] base     
#> 
#> other attached packages:
#>  [1] PharmacoGx_3.3.4            CoreGx_2.2.0               
#>  [3] SummarizedExperiment_1.28.0 Biobase_2.58.0             
#>  [5] GenomicRanges_1.50.2        GenomeInfoDb_1.34.9        
#>  [7] IRanges_2.32.0              S4Vectors_0.36.2           
#>  [9] MatrixGenerics_1.10.0       matrixStats_0.63.0         
#> [11] BiocGenerics_0.44.0        
#> 
#> loaded via a namespace (and not attached):
#>   [1] fgsea_1.24.0                colorspace_2.1-0           
#>   [3] ellipsis_0.3.2              rprojroot_2.0.3            
#>   [5] lsa_0.73.3                  XVector_0.38.0             
#>   [7] fs_1.6.1                    farver_2.1.1               
#>   [9] SnowballC_0.7.0             MultiAssayExperiment_1.24.0
#>  [11] DT_0.27                     fansi_1.0.4                
#>  [13] xml2_1.3.3                  codetools_0.2-18           
#>  [15] downlit_0.4.2               cachem_1.0.7               
#>  [17] knitr_1.42                  jsonlite_1.8.4             
#>  [19] magicaxis_2.2.14            cluster_2.1.3              
#>  [21] shinydashboard_0.7.2        shiny_1.7.4                
#>  [23] mapproj_1.2.11              compiler_4.2.1             
#>  [25] httr_1.4.5                  backports_1.4.1            
#>  [27] Matrix_1.4-1                fastmap_1.1.1              
#>  [29] bench_1.1.2                 limma_3.54.2               
#>  [31] cli_3.6.1                   later_1.3.0                
#>  [33] visNetwork_2.1.2            htmltools_0.5.5            
#>  [35] tools_4.2.1                 igraph_1.4.2               
#>  [37] gtable_0.3.3                glue_1.6.2                 
#>  [39] GenomeInfoDbData_1.2.9      reshape2_1.4.4             
#>  [41] RANN_2.6.1                  dplyr_1.1.2                
#>  [43] maps_3.4.1                  fastmatch_1.1-3            
#>  [45] Rcpp_1.0.10                 slam_0.1-50                
#>  [47] jquerylib_0.1.4             pkgdown_2.0.7              
#>  [49] vctrs_0.6.2                 BumpyMatrix_1.6.0          
#>  [51] stringr_1.5.0               xfun_0.38                  
#>  [53] mime_0.12                   lifecycle_1.0.3            
#>  [55] gtools_3.9.4                MASS_7.3-57                
#>  [57] zlibbioc_1.44.0             scales_1.2.1               
#>  [59] ragg_1.2.5                  promises_1.2.0.1           
#>  [61] relations_0.6-13            parallel_4.2.1             
#>  [63] RColorBrewer_1.1-3          sets_1.0-24                
#>  [65] yaml_2.3.7                  curl_5.0.0                 
#>  [67] memoise_2.0.1               ggplot2_3.4.2              
#>  [69] downloader_0.4              sass_0.4.5                 
#>  [71] stringi_1.7.12              NISTunits_1.0.1            
#>  [73] desc_1.4.2                  plotrix_3.8-2              
#>  [75] checkmate_2.1.0             caTools_1.18.2             
#>  [77] boot_1.3-28                 BiocParallel_1.32.6        
#>  [79] systemfonts_1.0.4           rlang_1.1.0                
#>  [81] pkgconfig_2.0.3             bitops_1.0-7               
#>  [83] pracma_2.4.2                evaluate_0.20              
#>  [85] lattice_0.20-45             purrr_1.0.1                
#>  [87] labeling_0.4.2              htmlwidgets_1.6.2          
#>  [89] cowplot_1.1.1               tidyselect_1.2.0           
#>  [91] plyr_1.8.8                  magrittr_2.0.3             
#>  [93] R6_2.5.1                    gplots_3.1.3               
#>  [95] generics_0.1.3              sm_2.2-5.7.1               
#>  [97] DelayedArray_0.24.0         pillar_1.9.0               
#>  [99] whisker_0.4.1               withr_2.5.0                
#> [101] RCurl_1.98-1.12             tibble_3.2.1               
#> [103] crayon_1.5.2                KernSmooth_2.23-20         
#> [105] utf8_1.2.3                  rmarkdown_2.21             
#> [107] grid_4.2.1                  data.table_1.14.8          
#> [109] marray_1.76.0               piano_2.14.0               
#> [111] digest_0.6.31               xtable_1.8-4               
#> [113] httpuv_1.6.9                BiocBaseUtils_1.0.0        
#> [115] textshaping_0.3.6           celestial_1.4.6            
#> [117] coop_0.6-3                  munsell_0.5.0              
#> [119] bslib_0.4.2                 shinyjs_2.1.0